refactor(orders): split ICE replay into diplomatic + prefix table (#4109)
Wave 5 slice C extracts diplomatic prefix replay and shared economy
projection probes, and drives recruit/build acceptance through
ProjectedResourcePrefixReplayConfig.

Refs #4109
